pyproject.toml
[project]
name = "dragonfly-reporter"
version = "0.1.0"
description = "Reports malicious packages to PyPI."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.12.dev0"
dependencies = [
  "fastapi-slim==0.115.2",
  "httpx==0.27.2",
  "pydantic-settings==2.6.0",
  "uvicorn[standard]==0.32.0",
  "sentry-sdk[fastapi]==2.17.0",
]

  [project.optional-dependencies]
  dev = ["pre-commit==4.0.1", "pyright==1.1.385", "ruff==0.7.0"]
  test = ["pytest==8.3.3"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
format = "ruff format"
lint = "ruff check --fix"
type-check = "pyright"
pre-commit = "pre-commit run --all-files"
test = "pytest"
all = {composite = ["format", "lint", "type-check", "test"]}

[tool.pyright]
include = ["src/**/*.py"]

venvPath = "."
venv = ".venv"

[tool.ruff]
preview = false
unsafe-fixes = true
target-version = "py311"
line-length = 120

  [tool.ruff.lint]
  select = ["ALL"]
  ignore = [
    "CPY001",  # (Missing copyright notice at top of file)
    "G004",    # (Logging statement uses f-string) - Developer UX
    "S101",    # (Use of `assert` detected) - This should probably be changed
    "PLR6301", # (Method `x` could be a function, class method, or static method) - false positives
    "ANN101",  # (Missing type annotation for `cls` in method) - Also for `self`; this just adds noise
  ]

    [tool.ruff.lint.extend-per-file-ignores]
    "docs/*" = [
      "INP001", # (File `docs/*.py` is part of an implicit namespace package. Add an `__init__.py`.) - Docs are not modules
    ]
    "tests/*" = [
      "INP001", # (File `tests/*.py` is part of an implicit namespace package. Add an `__init__.py`.) - Tests are not modules
      "S101",   # (Use of `assert` detected) - Yes, that's the point
    ]

    [tool.ruff.lint.pydocstyle]
    convention = "numpy"
